Refactor moving to Core directory structure and baseUrl set to src
