+++ src/core/activityPub/activityPub.ts
@@ -1,5 +1,5 @@
 import { keccak256 } from "js-sha3";
-import { HexString } from "../types/Strings";
+import { HexString } from "types/Strings";
 import { sortObject } from "../utilities/json";
 
 const ISO8601_REGEX = /(\d{4})-(\d{2})-(\d{2})T(\d{2}):(\d{2}):(\d{2})([+-](\d{2}):(\d{2}))?/;
