+++ src/core/utilities/guards.ts
