+++ src/core/utilities/hash.ts
@@ -1,4 +1,4 @@
-import { HexString } from "../types/Strings";
+import { HexString } from "types/Strings";
 
 export const hashPrefix = (hash: string): HexString => {
   return hash.startsWith("0x", 0) ? hash : `0x${hash}`;
