+++ src/index.ts
@@ -9,29 +9,12 @@
  *   - `import foo from "@dsnp/sdk/foo";`
  */
 
-import * as configImport from "./config/config";
-export const config = configImport;
+import * as coreImport from "./core";
+export const core = coreImport;
 
-import * as contractsImport from "./contracts";
-export const contracts = contractsImport;
-
-import * as activityPubImport from "./activityPub";
-export const activityPub = activityPubImport;
-
-import * as queueImport from "./queue";
-export const queue = queueImport;
-
-import * as utilitiesImport from "./utilities";
-export const utilities = utilitiesImport;
-
-import * as socialContentImport from "./social/content";
-export const socialContent = socialContentImport;
-
-import * as socialHandlesImport from "./social/handles";
-export const socialHandles = socialHandlesImport;
-
-import * as socialNetworkImport from "./social/network";
-export const socialNetwork = socialNetworkImport;
-
-import * as socialSearchImport from "./social/search";
-export const socialSearch = socialSearchImport;
+// Porcelain
+export * from "./config";
+export * from "./content";
+export * from "./handles";
+export * from "./network";
+export * from "./search";
