+++ src/test/generators/batchFileGenerators.ts
@@ -5,9 +5,9 @@ const path = require("path");
 const parquet = require("@dsnp/parquetjs");
 import { times } from "lodash";
 import { generateBroadcast, generateReply, generateReaction } from "./dsnpGenerators";
-import * as pq from "../../batch/parquetSchema";
-import { EthereumAddress } from "../../types/Strings";
-import { BatchBroadcastMessage, BatchReactionMessage, BatchReplyMessage } from "../../batch/batchMesssages";
+import * as pq from "core/batch/parquetSchema";
+import { EthereumAddress } from "types/Strings";
+import { BatchBroadcastMessage, BatchReactionMessage, BatchReplyMessage } from "core/batch/batchMesssages";
 import { generateHexString } from "@dsnp/test-generators";
 
 export type DSNPBatchWriteResult = {
