+++ src/test/generators/dsnpGenerators.ts
@@ -1,8 +1,8 @@
 import * as fs from "fs";
 import { times } from "lodash";
 
-import { BroadcastMessage, DSNPMessage, DSNPType, ReactionMessage, ReplyMessage } from "../../messages/messages";
-import { EthereumAddress } from "../../types/Strings";
+import { BroadcastMessage, DSNPMessage, DSNPType, ReactionMessage, ReplyMessage } from "core/messages/messages";
+import { EthereumAddress } from "types/Strings";
 import { generateHexString, randInt, sample } from "@dsnp/test-generators";
 import { addresses, sampleText } from "@dsnp/test-generators/";
