+++ tsconfig.json
@@ -3,6 +3,7 @@
     "target": "es6",
     "outDir": "dist/cjs",
     "rootDir": "src",
+    "baseUrl": "src",
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "inlineSourceMap": true,
